WebSocket controller for establishing authenticated connections.
Before a connection is allowed users must authenticate themeselves by
using the Heat decentralized authentication scheme where you sign a
message client side which the server app checks. The valid account id
will be bound to the connection and in case of abuse we can throttle it
or kill it.